+++ src/libopensc/cardctl.h
@@ -341,7 +341,7 @@ struct sc_cardctl_cardos_obj_info {
 
 struct sc_cardctl_cardos_genkey_info {
 	unsigned int	key_id;
-	unsigned int	key_bits;
+	size_t	key_bits;
 	unsigned short	fid;
 };
 
@@ -359,7 +359,7 @@ struct sc_cardctl_cardos_pass_algo_flags {
  */
 struct sc_cardctl_cryptoflex_genkey_info {
 	unsigned int	key_num;
-	unsigned int	key_bits;
+	size_t	key_bits;
 	unsigned long	exponent;
 	unsigned char *	pubkey;
 	unsigned int	pubkey_len;
@@ -417,10 +417,10 @@ enum SC_CARDCTL_OBERTHUR_KEY_TYPE {
 
 struct sc_cardctl_oberthur_genkey_info {
 	unsigned int    id_prv, id_pub;
-	unsigned int    key_bits;
+	size_t    key_bits;
 	unsigned long   exponent;
 	unsigned char * pubkey;
-	unsigned int    pubkey_len;
+	size_t    pubkey_len;
 
 	int     method;     /* SC_AC_XXX */
 	int     key_ref;    /* key reference */
@@ -433,14 +433,14 @@ struct sc_cardctl_oberthur_updatekey_info {
 	unsigned int    data_len;
 
 	unsigned char   id[256];
-	unsigned int    id_len;
+	size_t    id_len;
 };
 
 struct sc_cardctl_oberthur_createpin_info {
 	unsigned int type;
 	unsigned int ref;
 	const unsigned char *pin;
-	unsigned int pin_len;
+	size_t pin_len;
 	unsigned int pin_tries;
 	const unsigned char *puk;
 	unsigned int puk_len;
@@ -460,12 +460,12 @@ struct sc_cardctl_setcos_data_obj {
 
 struct sc_cardctl_setcos_gen_store_key_info {
 	int             op_type;
-	unsigned int    mod_len;     /* in bits */
-	unsigned int    pubexp_len;  /* in bits */
+	size_t    mod_len;     /* in bits */
+	size_t    pubexp_len;  /* in bits */
 	unsigned char  *pubexp;
-	unsigned int    primep_len;  /* in bits */
+	size_t    primep_len;  /* in bits */
 	unsigned char  *primep;
-	unsigned int    primeq_len;  /* in bits */
+	size_t    primeq_len;  /* in bits */
 	unsigned char  *primeq;
 };
 
@@ -474,7 +474,7 @@ struct sc_cardctl_setcos_gen_store_key_info {
  */
 typedef struct sc_cardctl_muscle_gen_key_info {
 	int 	keyType;
-	int 	keySize;
+	size_t 	keySize;
 	int 	privateKeyLocation;
 	int 	publicKeyLocation;
 } sc_cardctl_muscle_gen_key_info_t;
@@ -483,7 +483,7 @@ typedef struct sc_cardctl_muscle_gen_key_info {
 typedef struct sc_cardctl_muscle_key_info {
 	int 	keyType;
 	int 	keyLocation;
-	int 	keySize;
+	size_t 	keySize;
 	size_t 	modLength;
 	u8* 	modValue;
 	size_t 	expLength;
@@ -760,7 +760,7 @@ typedef struct sc_epass2003_gen_key_data_st {
 	 int pukey_id;
 	 size_t key_length;
 	 u8 *modulus;
-	 int modulus_len;
+	 size_t modulus_len;
 } sc_epass2003_gen_key_data;
 
 
@@ -847,13 +847,13 @@ typedef struct sc_cardctl_piv_genkey_info_st {
 	unsigned int	key_algid;	/* RSA 5, 6, 7; EC 11, 14 */
 	unsigned int	key_bits;	/* RSA */
 	unsigned char *	exponent;	/* RSA */
-	unsigned int	exponent_len;	/* RSA */
+	size_t	exponent_len;	/* RSA */
 	unsigned char *	pubkey;		/* RSA */
-	unsigned int	pubkey_len;	/* RSA */
+	size_t	pubkey_len;	/* RSA */
 	unsigned char * ecparam;        /* EC */
 	unsigned int    ecparam_len;    /* EC */
 	unsigned char * ecpoint;        /* EC */
-	unsigned int    ecpoint_len;    /* EC */
+	size_t    ecpoint_len;    /* EC */
 
 } sc_cardctl_piv_genkey_info_t;
 
@@ -896,7 +896,7 @@ typedef struct sc_cardctl_openpgp_keygen_info {
 			size_t ecpoint_len;
 			struct sc_object_id oid;
 			u8 oid_len;
-			unsigned int key_length;
+			size_t key_length;
 		} ec;
 	} u;
 } sc_cardctl_openpgp_keygen_info_t;
