+++ src/libopensc/cwa-dnie.c
@@ -718,8 +718,8 @@ static int dnie_get_root_ca_pubkey(sc_card_t * card, EVP_PKEY ** root_ca_key)
 	LOG_TEST_RET(card->ctx, res, "Error getting the card channel data");
 
 	/* compose root_ca_public key with data provided by Dnie Manual */
-	root_ca_rsa_n = BN_bin2bn(data->icc_root_ca.modulus.value, data->icc_root_ca.modulus.len, NULL);
-	root_ca_rsa_e = BN_bin2bn(data->icc_root_ca.exponent.value, data->icc_root_ca.exponent.len, NULL);
+	root_ca_rsa_n = BN_bin2bn(data->icc_root_ca.modulus.value, (int)data->icc_root_ca.modulus.len, NULL);
+	root_ca_rsa_e = BN_bin2bn(data->icc_root_ca.exponent.value, (int)data->icc_root_ca.exponent.len, NULL);
 
 #if OPENSSL_VERSION_NUMBER < 0x30000000L
 	if (RSA_set0_key(root_ca_rsa, root_ca_rsa_n, root_ca_rsa_e, NULL) != 1) {
@@ -871,8 +871,8 @@ static int dnie_get_cvc_ifd_cert_pin(sc_card_t * card, u8 ** cert, size_t * leng
  */
 static int dnie_get_privkey(sc_card_t * card, EVP_PKEY ** ifd_privkey,
                             u8 * modulus, int modulus_len,
-                            u8 * public_exponent, int public_exponent_len,
-                            u8 * private_exponent, int private_exponent_len)
+                            u8 * public_exponent, size_t public_exponent_len,
+                            u8 * private_exponent, size_t private_exponent_len)
 {
 	BIGNUM *ifd_rsa_n = NULL, *ifd_rsa_e = NULL, *ifd_rsa_d = NULL;
 
@@ -905,8 +905,8 @@ static int dnie_get_privkey(sc_card_t * card, EVP_PKEY ** ifd_privkey,
 
 	/* compose ifd_private key with data provided in Annex 3 of DNIe Manual */
 	ifd_rsa_n = BN_bin2bn(modulus, modulus_len, NULL);
-	ifd_rsa_e = BN_bin2bn(public_exponent, public_exponent_len, NULL);
-	ifd_rsa_d = BN_bin2bn(private_exponent, private_exponent_len, NULL);
+	ifd_rsa_e = BN_bin2bn(public_exponent, (int)public_exponent_len, NULL);
+	ifd_rsa_d = BN_bin2bn(private_exponent, (int)private_exponent_len, NULL);
 
 #if OPENSSL_VERSION_NUMBER < 0x30000000L
 	if (RSA_set0_key(ifd_rsa, ifd_rsa_n, ifd_rsa_e, ifd_rsa_d) != 1) {
@@ -980,7 +980,7 @@ static int dnie_get_ifd_privkey(sc_card_t * card, EVP_PKEY ** ifd_privkey)
 	res = dnie_get_channel_data(card, &data);
 	LOG_TEST_RET(card->ctx, res, "Error getting the card channel data");
 
-	return dnie_get_privkey(card, ifd_privkey, data->ifd.modulus.value, data->ifd.modulus.len,
+	return dnie_get_privkey(card, ifd_privkey, data->ifd.modulus.value, (int)data->ifd.modulus.len,
 				data->ifd.exponent.value, data->ifd.exponent.len,
 				data->ifd.private.value, data->ifd.private.len);
 }
@@ -1004,7 +1004,7 @@ static int dnie_get_ifd_privkey_pin(sc_card_t * card, EVP_PKEY ** ifd_privkey)
 	res = dnie_get_channel_data(card, &data);
 	LOG_TEST_RET(card->ctx, res, "Error getting the card channel data");
 
-	return dnie_get_privkey(card, ifd_privkey, data->ifd_pin.modulus.value, data->ifd_pin.modulus.len,
+	return dnie_get_privkey(card, ifd_privkey, data->ifd_pin.modulus.value, (int)data->ifd_pin.modulus.len,
 				data->ifd_pin.exponent.value, data->ifd_pin.exponent.len,
 				data->ifd_pin.private.value, data->ifd_pin.private.len);
 }
@@ -1355,7 +1355,7 @@ void dnie_change_cwa_provider_to_pin(sc_card_t * card)
 }
 
 void dnie_format_apdu(sc_card_t *card, sc_apdu_t *apdu,
-			int cse, int ins, int p1, int p2, int le, int lc,
+			int cse, int ins, int p1, int p2, size_t le, size_t lc,
 			unsigned char * resp, size_t resplen,
 			const unsigned char * data, size_t datalen)
 {
