+++ src/libopensc/opensc.h
@@ -238,16 +238,16 @@ struct sc_supported_algo_info {
 typedef struct sc_sec_env_param {
 	unsigned int param_type;
 	void* value;
-	unsigned int value_len;
+	size_t value_len;
 } sc_sec_env_param_t;
 
 
 typedef struct sc_security_env {
 	unsigned long flags;
 	int operation;
-	unsigned int algorithm, algorithm_flags;
+	unsigned long algorithm, algorithm_flags;
 
-	unsigned int algorithm_ref;
+	unsigned long algorithm_ref;
 	struct sc_path file_ref;
 	unsigned char key_ref[8];
 	size_t key_ref_len;
@@ -259,7 +259,7 @@ typedef struct sc_security_env {
 } sc_security_env_t;
 
 struct sc_algorithm_id {
-	unsigned int algorithm;
+	unsigned long algorithm;
 	struct sc_object_id oid;
 	void *params;
 };
@@ -298,15 +298,15 @@ struct sc_ec_parameters {
 
 typedef struct sc_algorithm_info {
 	unsigned int algorithm;
-	unsigned int key_length;
-	unsigned int flags;
+	size_t key_length;
+	unsigned long flags;
 
 	union {
 		struct sc_rsa_info {
 			unsigned long exponent;
 		} _rsa;
 		struct sc_ec_info {
-			unsigned ext_flags;
+			unsigned long ext_flags;
 			struct sc_ec_parameters params;
 		} _ec;
 	} u;
@@ -444,7 +444,7 @@ struct sc_pin_cmd_pin {
 	const char *prompt;	/* Prompt to display */
 
 	const unsigned char *data; /* PIN, set to NULL when using pin pad */
-	int len;		/* set to 0 when using pin pad */
+	size_t len;		/* set to 0 when using pin pad */
 
 	size_t min_length;	/* min length of PIN */
 	size_t max_length;	/* max length of PIN */
@@ -1604,17 +1604,17 @@ void sc_invalidate_cache(struct sc_card *card);
 void sc_print_cache(struct sc_card *card);
 
 struct sc_algorithm_info * sc_card_find_rsa_alg(struct sc_card *card,
-		unsigned int key_length);
+		size_t key_length);
 struct sc_algorithm_info * sc_card_find_ec_alg(struct sc_card *card,
-		unsigned int field_length, struct sc_object_id *curve_oid);
+		size_t field_length, struct sc_object_id *curve_oid);
 struct sc_algorithm_info * sc_card_find_eddsa_alg(struct sc_card *card,
-		unsigned int field_length, struct sc_object_id *curve_oid);
+		size_t field_length, struct sc_object_id *curve_oid);
 struct sc_algorithm_info * sc_card_find_xeddsa_alg(struct sc_card *card,
-		unsigned int field_length, struct sc_object_id *curve_oid);
+		size_t field_length, struct sc_object_id *curve_oid);
 struct sc_algorithm_info * sc_card_find_gostr3410_alg(struct sc_card *card,
-		unsigned int key_length);
+		size_t key_length);
 struct sc_algorithm_info * sc_card_find_alg(sc_card_t *card,
-		unsigned int algorithm, unsigned int key_length, void *param);
+		unsigned int algorithm, size_t key_length, void *param);
 
 scconf_block *sc_match_atr_block(sc_context_t *ctx, struct sc_card_driver *driver, struct sc_atr *atr);
 /**
